+++ client/src/lib/APIs/reviewApi.ts
@@ -3,7 +3,7 @@ import { handleApiError } from "./userDataApi";
 
 const backendUrl = import.meta.env.VITE_BACKEND_URL;
 
-export const postReview = async (reviewData: Review) => {
+export const postReview = async (reviewData: Review, imageFormData: FormData | null = null) => {
     try {
         const response = await fetch(`${backendUrl}/api/review`, {
             method: "POST",
@@ -16,7 +16,33 @@ export const postReview = async (reviewData: Review) => {
         }
 
         const createdReview: Review = await response.json();
-        return createdReview;
+
+        if (imageFormData && createdReview.id) {
+            imageFormData.append("review_id", createdReview.id);
+
+            await postImages(imageFormData);
+        }
+
+        // return createdReview;
+    }
+    catch (error) {
+        handleApiError(error);
+    }
+}
+
+export const postImages = async (formData: FormData) => {
+    try {
+        const response = await fetch(`${backendUrl}/api/review/image`, {
+            method: "POST",
+            body: formData
+        });
+
+        if (!response.ok) {
+            throw new Error(`[Backend] Network error: ${response.status}`);
+        }
+
+        const imgUrlList: string[] = await response.json();
+        return imgUrlList;
     }
     catch (error) {
         handleApiError(error);
@@ -50,7 +76,7 @@ export const fetchReviewsWithUserDataByParkId = async (
 
 const deleteReviewImage = async (reviewId: string) => {
     try {
-        const response = await fetch(`${backendUrl}/api/reviewimage/reviewid/${reviewId}`, {
+        const response = await fetch(`${backendUrl}/api/review/image/review_id/${reviewId}`, {
             method: "DELETE",
         });
