+++ server/src/routers/reviewRouter.ts
@@ -1,7 +1,8 @@
 import { Request, Response, Router } from "express";
 import pool from "../db";
 import { handleError } from "../../middleware/errorHandler";
-import { deleteImage } from "../cloud/bucketFileManager";
+import { deleteImage, uploadImage } from "../cloud/bucketFileManager";
+import { upload } from "../multer/multer";
 
 export const reviewRouter = Router();
 
@@ -9,37 +10,52 @@ reviewRouter.post("/", async (req: Request, res: Response) => {
     try {
         const { user_sub_id, park_id, rating, content } = req.body;
 
-        // let insertQueryResult;
-        // if (req.files) {
-        //     const fileList = req.files as Express.Multer.File[];
-        //     let img_url_list: string[] = [];
-
-        //     fileList.forEach(async (file: Express.Multer.File) => {
-        //         const img_url = await uploadImage(file)
-        //         img_url_list.push(img_url);
-        //     });
-        //     console.log(img_url_list);
-
-        //     insertQueryResult = await pool.query(
-        //         "INSERT INTO Review(user_sub_id, park_id, rating, content, img_url) VALUES($1, $2, $3, $4, $5) RETURNING *",
-        //         [user_sub_id, park_id, rating, content, img_url_list]
-        //     );
-        // }
-        // // Refactor below
-        // else if (req.file) {
-        //     const img_url = await uploadImage(req.file)
-
-        //     insertQueryResult = await pool.query(
-        //         "INSERT INTO Review(user_sub_id, park_id, rating, content, img_url) VALUES($1, $2, $3, $4, $5) RETURNING *",
-        //         [user_sub_id, park_id, rating, content, img_url]
-        //     );
-        // }
         const insertQueryResult = await pool.query(
             "INSERT INTO Review(user_sub_id, park_id, rating, content) VALUES($1, $2, $3, $4) RETURNING *",
             [user_sub_id, park_id, rating, content]
         );
 
         res.json(insertQueryResult.rows[0]);
+    } catch (error) {
+        handleError(error, res);
+    }
+});
+
+reviewRouter.post("/image", upload.array("img_file"), async (req: Request, res: Response) => {
+    try {
+        if (!req.files) {
+            throw new Error(
+                "[Backend] Invalid Data: the image list is empty!"
+            );
+        }
+
+        const { review_id } = req.body;
+        const fileList = req.files as Express.Multer.File[];
+
+        // Apparently foreach or map functions don't wait for async functions
+        // You can either use Promise.all or for loop (for more control within the loop).
+        // explicit return statement in .map() is omitted since async functions implicitly returns promises
+        await Promise.all(
+            fileList.map(async (file) => {
+                const img_url = await uploadImage(file);
+                console.log(img_url);
+
+                await pool.query(
+                    "INSERT INTO ReviewImage(review_id, img_url) VALUES ($1, $2)",
+                    [review_id, img_url]
+                );
+            })
+        );
+
+        const selectQueryResult = await pool.query(
+            "SELECT img_url FROM ReviewImage WHERE review_id=$1",
+            [review_id]
+        );
+        const imgUrlList = selectQueryResult.rows.map((reviewImage) => {
+            return reviewImage["img_url"];
+        });
+
+        res.json(imgUrlList);
     } catch (error) {
         handleError(error, res);
     }
@@ -70,26 +86,10 @@ reviewRouter.get("/park_id/:park_id", async (req: Request, res: Response) => {
     }
 });
 
-interface test {
-    list: string[];
-}
-
-reviewRouter.get("/test", async (req: Request, res: Response) => {
-    try {
-        const x: test = {
-            list: ["a", "b"],
-        };
-
-        res.json(x);
-    } catch (error) {
-        handleError(error, res);
-    }
-});
-
 reviewRouter.get("/reviewWithUserData", async (req: Request, res: Response) => {
     try {
         const columns =
-            "Review.id AS review_id, Review.park_id, Review.rating, Review.content, Review.created_at, Review.user_sub_id, UserData.name AS user_name, UserData.profile_picture_url AS user_profile_picture_url, ThumbsUpList.id AS thumbs_up_id";
+            "Review.id AS review_id, Review.park_id, Review.rating, Review.content, Review.created_at, Review.user_sub_id, UserData.name AS user_name, UserData.profile_picture_url AS user_profile_picture_url, ThumbsUpList.id AS thumbs_up_id, (SELECT ARRAY_AGG(img_url) AS img_url_list FROM ReviewImage WHERE ReviewImage.review_id=Review.id)";
 
         const { park_id, user_sub_id = "" } = req.query;
 
@@ -97,6 +97,8 @@ reviewRouter.get("/reviewWithUserData", async (req: Request, res: Response) => {
             `SELECT ${columns} FROM REVIEW LEFT JOIN ThumbsUpList ON Review.id=ThumbsUpList.review_id AND ThumbsUpList.user_sub_id=$1 JOIN UserData ON Review.user_sub_id=UserData.sub_id WHERE park_id=$2 ORDER BY Review.created_at DESC`,
             [user_sub_id, park_id]
         );
+
+        // now it returns urls as list
 
         res.json(selectQueryResult.rows);
     } catch (error) {
@@ -137,6 +139,38 @@ reviewRouter.get("/review_id/:id", async (req: Request, res: Response) => {
     }
 });
 
+reviewRouter.get("/image/:id", async (req: Request, res: Response) => {
+    try {
+        const { id } = req.params;
+
+        const selectQueryResult = await pool.query(
+            "SELECT * FROM ReviewImage WHERE id=$1",
+            [id]
+        );
+
+        res.json(selectQueryResult.rows[0]);
+    } catch (error) {
+        handleError(error, res);
+    }
+});
+
+reviewRouter.get("/image/review_id/:review_id",
+    async (req: Request, res: Response) => {
+        try {
+            const { review_id } = req.params;
+
+            const selectQueryResult = await pool.query(
+                "SELECT * FROM ReviewImage WHERE review_id=$1",
+                [review_id]
+            );
+
+            res.json(selectQueryResult.rows);
+        } catch (error) {
+            handleError(error, res);
+        }
+    }
+);
+
 reviewRouter.put("/review_id/:id", async (req: Request, res: Response) => {
     try {
         const { id } = req.params;
@@ -167,3 +201,55 @@ reviewRouter.delete("/review_id/:id", async (req: Request, res: Response) => {
         handleError(error, res);
     }
 });
+
+reviewRouter.delete("/image/:id", async (req: Request, res: Response) => {
+    try {
+        const { id } = req.params;
+
+        const selectImgUrlQueryResult = await pool.query(
+            "SELECT * FROM ReviewImage WHERE id=$1",
+            [id]
+        );
+        const img_url = selectImgUrlQueryResult.rows[0]["img_url"];
+        await deleteImage(img_url);
+
+        const deleteQueryResult = await pool.query(
+            "DELETE FROM ReviewImage WHERE id=$1",
+            [id]
+        );
+
+        res.json(deleteQueryResult);
+    } catch (error) {
+        handleError(error, res);
+    }
+});
+
+reviewRouter.delete("/image/review_id/:review_id",
+    async (req: Request, res: Response) => {
+        try {
+            const { review_id } = req.params;
+
+            const selectQueryResult = await pool.query(
+                "SELECT * FROM ReviewImage WHERE review_id=$1",
+                [review_id]
+            );
+
+            await Promise.all(
+                selectQueryResult.rows.map(async (reviewImage) => {
+                    const img_url = reviewImage["img_url"];
+
+                    await deleteImage(img_url);
+                })
+            );
+
+            const deleteQueryResult = await pool.query(
+                "DELETE FROM ReviewImage WHERE review_id=$1",
+                [review_id]
+            );
+
+            res.json(deleteQueryResult);
+        } catch (error) {
+            handleError(error, res);
+        }
+    }
+);
